backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.1",
    "fastapi-socketio>=0.0.10",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "motor>=3.7.1",
    "openai>=2.16.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "python-socketio>=5.16.0",
    "redis>=7.1.0",
    "slowapi>=0.1.9",
    "uvicorn>=0.40.0",
]
